requirements.txt
python-telegram-bot==20.7
steampy==1.2.0
requests==2.31.0
httpx
python-dotenv==1.0.0
APScheduler==3.10.4
aiohttp==3.9.3
